ref(dspy): migrate dspy wrapper to integrations API
Move the DSPy provider from braintrust.wrappers.dspy into
braintrust.integrations.dspy following the standard integration
package layout (integration.py, patchers.py, tracing.py).

- DSPyIntegration uses FunctionWrapperPatcher to patch dspy.configure
- BraintrustDSpyCallback and _configure_wrapper live in tracing.py
- Wrapper module reduced to compatibility re-exports
- auto.py uses DSPyIntegration.setup() instead of the old wrapper
- Tests, cassettes, and auto-test script moved to integration dir
- noxfile updated to point at new test location

FunctionWrapperPatcher.mark_patched/is_patched/patch now handle
targets that don't support setattr (e.g. bound methods) by falling
back to storing the marker on the root module. This avoids the need
for per-patcher overrides when the patch target is a bound method,
as is the case with dspy.configure.
