+++ py/src/braintrust/integrations/__init__.py
@@ -2,6 +2,7 @@ from .adk import ADKIntegration
 from .agno import AgnoIntegration
 from .anthropic import AnthropicIntegration
 from .claude_agent_sdk import ClaudeAgentSDKIntegration
+from .dspy import DSPyIntegration
 from .google_genai import GoogleGenAIIntegration
 
 
@@ -10,5 +11,6 @@ __all__ = [
     "AgnoIntegration",
     "AnthropicIntegration",
     "ClaudeAgentSDKIntegration",
+    "DSPyIntegration",
     "GoogleGenAIIntegration",
 ]
