+++ py/src/braintrust/integrations/dspy/__init__.py
@@ -1,0 +1,12 @@
+"""Braintrust integration for DSPy."""
+
+from .integration import DSPyIntegration
+from .patchers import patch_dspy
+from .tracing import BraintrustDSpyCallback
+
+
+__all__ = [
+    "BraintrustDSpyCallback",
+    "DSPyIntegration",
+    "patch_dspy",
+]
