+++ py/src/braintrust/integrations/dspy/integration.py
@@ -1,0 +1,13 @@
+"""DSPy integration — orchestration class and setup entry-point."""
+
+from braintrust.integrations.base import BaseIntegration
+
+from .patchers import DSPyConfigurePatcher
+
+
+class DSPyIntegration(BaseIntegration):
+    """Braintrust instrumentation for DSPy."""
+
+    name = "dspy"
+    import_names = ("dspy",)
+    patchers = (DSPyConfigurePatcher,)
